+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.1.13](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.1.12...@stream-io/video-react-native-sdk-0.1.13) (2023-10-25)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.3.36`
+* `@stream-io/video-react-bindings` updated to version `0.2.37`
 ### [0.1.12](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.1.11...@stream-io/video-react-native-sdk-0.1.12) (2023-10-20)